creating mirrored structure
