+++ pytensor/link/jax/dispatch/__init__.py
@@ -11,6 +11,7 @@ import pytensor.link.jax.dispatch.pad
 import pytensor.link.jax.dispatch.math
 import pytensor.link.jax.dispatch.linalg
 import pytensor.link.jax.dispatch.random
+import pytensor.link.jax.dispatch.reshape
 import pytensor.link.jax.dispatch.scalar
 import pytensor.link.jax.dispatch.scan
 import pytensor.link.jax.dispatch.shape
