+++ pytensor/link/jax/dispatch/shape.py
@@ -4,10 +4,6 @@ from pytensor.graph import Constant
 from pytensor.graph.basic import Apply
 from pytensor.graph.op import Op
 from pytensor.link.jax.dispatch.basic import jax_funcify
-from pytensor.tensor.reshape import (  # is the placement of joindims/splitdims under reshape instead of shape intentional?
-    JoinDims,
-    SplitDims,
-)
 from pytensor.tensor.shape import Reshape, Shape, Shape_i, SpecifyShape
 from pytensor.tensor.type import TensorType
 
@@ -108,26 +104,3 @@ def jax_funcify_SpecifyShape(op, node, **kwargs):
         return x
 
     return specifyshape
-
-
-@jax_funcify.register(JoinDims)
-def jax_funcify_JoinDims(op, node, **kwargs):
-    start_axis = op.start_axis
-    n_axes = op.n_axes
-
-    def join_dims(x):
-        output_shape = (*x.shape[:start_axis], -1, *x.shape[start_axis + n_axes :])
-        return jnp.reshape(x, output_shape)
-
-    return join_dims
-
-
-@jax_funcify.register(SplitDims)
-def jax_funcify_SplitDims(op, node, **kwargs):
-    axis = op.axis
-
-    def split_dims(x, shape):
-        output_shape = (*x.shape[:axis], *shape, *x.shape[axis + 1 :])
-        return jnp.reshape(x, output_shape)
-
-    return split_dims
