+++ pytensor/link/numba/dispatch/__init__.py
@@ -8,6 +8,7 @@ import pytensor.link.numba.dispatch.elemwise
 import pytensor.link.numba.dispatch.extra_ops
 import pytensor.link.numba.dispatch.linalg
 import pytensor.link.numba.dispatch.random
+import pytensor.link.numba.dispatch.reshape
 import pytensor.link.numba.dispatch.scan
 import pytensor.link.numba.dispatch.scalar
 import pytensor.link.numba.dispatch.shape
